% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{By using run_app() function in 'TestAnaAPP' package without any arguments,
 you can perform your test analysis in a friendly interactive interface.}
\usage{
run_app()
}
\value{
No explicit return value. The function is called for its side effects,
which include running the interactive application.
}
\description{
By using run_app() function in 'TestAnaAPP' package without any arguments,
 you can perform your test analysis in a friendly interactive interface.
}
\examples{
if(interactive()){
TestAnaAPP::run_app()
}

}
